Extract writeToPdfFile to StorageOperations method
